Fix minimap highlighting
Now, even if the file is very large, the highlights will
be correctly rendered for parts of the file that are not in
currently displayed portion of the minimap. Make use of the
offset, in the form of startLineNumber when the minimap is
too long. Also, if you scroll using the minimap, the
highlights won't go away.

Signed-off-by: Abhijit Kiran Valluri <me@abhijitvalluri.com>